requirements.txt
Flask==2.3.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
numpy==1.24.3
psycopg2-binary==2.9.7
pgvector==0.2.4